+++ contracts/ClearingHouse.sol
@@ -103,6 +103,7 @@ contract ClearingHouse is
     //
 
     /// @dev this function is public for testing
+    // solhint-disable-next-line func-order
     function initialize(
         address clearingHouseConfigArg,
         address vaultArg,
@@ -191,7 +192,7 @@ contract ClearingHouse is
         // price slippage check
         require(response.base >= params.minBase && response.quote >= params.minQuote, "CH_PSC");
 
-        // if useTakerBalance, use addTakerBalance() instead of addBalance() to modify takerBalances
+        // if !useTakerBalance, takerBalance won't change, only need to collects fee to oweRealizedPnl
         if (params.useTakerBalance) {
             bool isBaseAdded = response.base > 0;
             bool isQuoteAdded = response.quote > 0;
@@ -233,26 +234,11 @@ contract ClearingHouse is
             IOrderBook(_orderBook).updateOrderDebt(response.orderId, deltaBaseDebt, deltaQuoteDebt);
 
             // update takerBalances as we're using takerBalances to provide liquidity
-            IAccountBalance(_accountBalance).addTakerBalances(
-                trader,
-                params.baseToken,
-                response.base.neg256(),
-                response.quote.neg256(),
-                deltaBaseDebt,
-                deltaQuoteDebt,
-                response.fee.toInt256()
-            );
-        } else {
-            IAccountBalance(_accountBalance).addBalance(
-                trader,
-                params.baseToken,
-                response.base.neg256(),
-                response.quote.neg256(),
-                response.fee.toInt256()
-            );
+            IAccountBalance(_accountBalance).addTakerBalances(trader, params.baseToken, deltaBaseDebt, deltaQuoteDebt);
         }
 
         // fees always have to be collected to owedRealizedPnl, as long as there is a change in liquidity
+        IAccountBalance(_accountBalance).addOwedRealizedPnl(trader, response.fee.toInt256());
 
         // after token balances are updated, we can check if there is enough free collateral
         _requireEnoughFreeCollateral(trader);
@@ -605,6 +591,7 @@ contract ClearingHouse is
         int256 fundingPayment = IExchange(_exchange).getAllPendingFundingPayment(trader);
         (int256 owedRealizedPnl, int256 unrealizedPnl) =
             IAccountBalance(_accountBalance).getOwedAndUnrealizedPnl(trader);
+        // solhint-disable-next-line var-name-mixedcase
         int256 balanceX10_18 =
             SettlementTokenMath.parseSettlementToken(IVault(_vault).getBalance(trader), _settlementTokenDecimals);
 
@@ -658,8 +645,6 @@ contract ClearingHouse is
         IAccountBalance(_accountBalance).settleBalanceAndDeregister(
             maker,
             baseToken,
-            response.base.toInt256(),
-            response.quote.toInt256(),
             response.deltaTakerBase,
             response.deltaTakerQuote,
             pnlToBeRealized,