+++ contracts/interface/IOrderBook.sol
@@ -103,13 +103,29 @@ interface IOrderBook {
 
     function hasOrder(address trader, address[] calldata tokens) external view returns (bool);
 
+    // @audit suggest to remove, especially if we can remove AccountBalance.getNetQuoteBalance - @wraecca
     function getTotalQuoteAmountInPools(address trader, address[] calldata baseTokens) external view returns (uint256);
+
+    function getTotalQuoteBalance(address trader, address[] calldata baseTokens) external view returns (int256);
 
     function getTotalTokenAmountInPool(
         address trader,
         address baseToken,
         bool fetchBase
     ) external view returns (uint256 tokenAmount);
+
+    function getTotalOrderDebt(
+        address trader,
+        address baseToken,
+        bool fetchBase
+    ) external view returns (uint256);
+
+    // getMakerBalance = totalTokenAmountInPool - totalOrderDebt
+    function getMakerBalance(
+        address trader,
+        address baseToken,
+        bool fetchBase
+    ) external view returns (int256);
 
     function getLiquidityCoefficientInFundingPayment(
         address trader,