+++ contracts/lib/AccountMarket.sol
@@ -10,8 +10,6 @@ library AccountMarket {
 
     /// @param lastTwPremiumGrowthGlobalX96 the last time weighted premiumGrowthGlobalX96
     struct Info {
-        int256 baseBalance;
-        int256 quoteBalance;
         int256 takerBaseBalance;
         int256 takerQuoteBalance;
         int256 lastTwPremiumGrowthGlobalX96;